+++ packages/explorable-explanations/src/protectedAudience/modules/flow.js
@@ -16,7 +16,7 @@
 /**
  * Internal dependencies.
  */
-import app from '../app';
+import app from '../app.js';
 import config from '../config.js';
 
 /**
